backend/requirements.txt
requests==2.26.0
django==2.2.16
psycopg2-binary==2.8.6
python-dotenv==0.20.0
djangorestframework==3.12.4
requests==2.26.0